requirements.txt
apache-airflow-providers-amazon==7.1.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-redis==3.1.0
apache-airflow-providers-mysql==4.0.0
boto3==1.26.51
boto==2.49.0
botocore==1.29.51
celery==5.2.7
flower==1.2.0
gunicorn==20.1.0
kombu==5.2.4
imbalanced-learn==0.8.1
selenium
imblearn
pymysql
pandas==1.5.2
numpy==1.22.4
pyarrow==9.0.0
scikit-learn==1.1.0